requirements.txt
Flask>=2.0.0
mysql-connector-python>=8.0.0
Werkzeug>=2.0.0
SQLAlchemy>=1.4.0
tomli>=2.0.0
fpdf>=1.7.2
pandas>=1.3.0
xlsxwriter>=3.0.0
python-docx>=0.8.11
# Add other dependencies as needed, for example:
# flask-sqlalchemy==3.0.5
# flask-migrate==4.0.4
